requirements.txt
pillow==11.0.0
pyperclip==1.9.0
